homepage updated

+++ include/publish.py
@@ -23,8 +23,11 @@ if __name__ == "__main__":
         print 'Section ' + section + ' doesn\'t exist!'
         sys.exit(1)
 
-    if os.path.exists('published/' + section):
-        shutil.rmtree('published/' + section)
-    os.mkdir('published/' + section)
+    if section is not './' or \
+            section is not '' or \
+            section is not '.':
+        if os.path.exists('published/' + section):
+            shutil.rmtree('published/' + section)
+        os.mkdir('published/' + section)
 
     publish(section)
